flask/requirements.txt
click==7.1.2
Flask==1.1.2
flask-restx==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
uWSGI==2.0.19.1
Werkzeug==1.0.1
debugpy==1.0.0
python-dotenv==0.14.0
flask_script==2.0.6
uuid==1.30